move structs to new modules (no code changes)
