+++ openmls/src/framing/codec.rs
@@ -1,12 +1,17 @@
 use tls_codec::{Deserialize, Serialize, Size};
 
-use super::*;
+use super::{
+    mls_auth_content::MlsContentAuthData,
+    mls_content::{ContentType, MlsContent, MlsContentBody},
+    plaintext::{MembershipTag, MlsPlaintext},
+    *,
+};
 use std::io::{Read, Write};
 
 impl Deserialize for MlsPlaintext {
     fn tls_deserialize<R: Read>(bytes: &mut R) -> Result<Self, tls_codec::Error> {
         let content: MlsContent = MlsContent::tls_deserialize(bytes)?;
-        let auth = deserialize_content_auth_data(bytes, content.body.content_type())?;
+        let auth = MlsContentAuthData::deserialize(bytes, content.body.content_type())?;
         let membership_tag = if content.sender.is_member() {
             Some(MembershipTag::tls_deserialize(bytes)?)
         } else {
@@ -43,91 +48,6 @@ impl Serialize for MlsPlaintext {
     }
 }
 
-pub(super) fn serialize_plaintext_tbs<'a, W: Write>(
-    wire_format: WireFormat,
-    content: &MlsContent,
-    serialized_context: impl Into<Option<&'a [u8]>>,
-    buffer: &mut W,
-) -> Result<usize, tls_codec::Error> {
-    let mut written = wire_format.tls_serialize(buffer)?;
-    written += content.tls_serialize(buffer)?;
-    written += if let Some(serialized_context) = serialized_context.into() {
-        // Only members and new members joining via commit should have a context.
-        debug_assert!(matches!(
-            content.sender,
-            Sender::Member(_) | Sender::NewMemberCommit
-        ));
-        buffer.write(serialized_context)?
-    } else {
-        0
-    };
-    Ok(written)
-}
-
-impl Size for MlsContentTbs {
-    #[inline]
-    fn tls_serialized_len(&self) -> usize {
-        self.wire_format.tls_serialized_len()
-            + self.content.tls_serialized_len()
-            + if let Some(serialized_context) = &self.serialized_context {
-                serialized_context.tls_serialized_len()
-            } else {
-                0
-            }
-    }
-}
-
-impl Serialize for MlsContentTbs {
-    fn tls_serialize<W: Write>(&self, writer: &mut W) -> Result<usize, tls_codec::Error> {
-        serialize_plaintext_tbs(
-            self.wire_format,
-            &self.content,
-            self.serialized_context.as_deref(),
-            writer,
-        )
-    }
-}
-
-impl Size for MlsContentAuthData {
-    #[inline]
-    fn tls_serialized_len(&self) -> usize {
-        self.signature.tls_serialized_len()
-            + if let Some(confirmation_tag) = &self.confirmation_tag {
-                confirmation_tag.tls_serialized_len()
-            } else {
-                0
-            }
-    }
-}
-
-impl Serialize for MlsContentAuthData {
-    fn tls_serialize<W: Write>(&self, writer: &mut W) -> Result<usize, tls_codec::Error> {
-        let mut written = self.signature.tls_serialize(writer)?;
-        written += if let Some(confirmation_tag) = &self.confirmation_tag {
-            confirmation_tag.tls_serialize(writer)?
-        } else {
-            0
-        };
-        Ok(written)
-    }
-}
-
-fn deserialize_content_auth_data<R: Read>(
-    bytes: &mut R,
-    content_type: ContentType,
-) -> Result<MlsContentAuthData, tls_codec::Error> {
-    let signature = Signature::tls_deserialize(bytes)?;
-    let confirmation_tag = if matches!(content_type, ContentType::Commit) {
-        Some(ConfirmationTag::tls_deserialize(bytes)?)
-    } else {
-        None
-    };
-    Ok(MlsContentAuthData {
-        signature,
-        confirmation_tag,
-    })
-}
-
 impl Size for MlsCiphertextContent {
     fn tls_serialized_len(&self) -> usize {
         self.content.tls_serialized_len() +
@@ -162,7 +82,7 @@ pub(super) fn deserialize_ciphertext_content<R: Read>(
     content_type: ContentType,
 ) -> Result<MlsCiphertextContent, tls_codec::Error> {
     let content = MlsContentBody::deserialize_without_type(bytes, content_type)?;
-    let auth = deserialize_content_auth_data(bytes, content_type)?;
+    let auth = MlsContentAuthData::deserialize(bytes, content_type)?;
 
     let padding = {
         let mut buffer = Vec::new();
