+++ openmls/src/framing/validation.rs
@@ -44,7 +44,11 @@ use crate::{
     tree::sender_ratchet::SenderRatchetConfiguration,
 };
 
-use super::*;
+use super::{
+    mls_auth_content::{MlsAuthContent, VerifiableMlsAuthContent},
+    mls_content::ContentType,
+    *,
+};
 
 /// Intermediate message that can be constructed either from a plaintext message or from ciphertext message.
 /// If it it constructed from a ciphertext message, the ciphertext message is decrypted first.
