+++ openmls/src/group/tests/kat_transcripts.rs
@@ -9,8 +9,17 @@ use std::convert::TryFrom;
 use crate::test_utils::{read, write};
 
 use crate::{
-    ciphersuite::signable::*, credentials::*, framing::*, group::*, messages::*, schedule::*,
-    test_utils::*, versions::ProtocolVersion,
+    ciphersuite::signable::*,
+    credentials::*,
+    framing::{
+        mls_auth_content::{MlsAuthContent, VerifiableMlsAuthContent},
+        *,
+    },
+    group::*,
+    messages::*,
+    schedule::*,
+    test_utils::*,
+    versions::ProtocolVersion,
 };
 
 use openmls_rust_crypto::OpenMlsRustCrypto;
