+++ openmls/src/test_utils/test_framework/mod.rs
@@ -24,7 +24,7 @@
 use crate::{
     ciphersuite::{hash_ref::KeyPackageRef, *},
     credentials::*,
-    framing::*,
+    framing::{mls_content::ContentType, *},
     group::*,
     key_packages::*,
     messages::*,
