+++ openmls/src/tree/tests_and_kats/kats/kat_encryption.rs
@@ -189,7 +189,9 @@ fn build_handshake_messages(
 ) -> (Vec<u8>, Vec<u8>) {
     use tls_codec::Serialize;
 
-    use crate::{prelude_test::Secret, schedule::MembershipKey};
+    use crate::{
+        framing::mls_auth_content::MlsAuthContent, prelude_test::Secret, schedule::MembershipKey,
+    };
 
     let epoch = random_u64();
     group.context_mut().set_epoch(epoch.into());
@@ -250,7 +252,9 @@ fn build_application_messages(
 ) -> (Vec<u8>, Vec<u8>) {
     use tls_codec::Serialize;
 
-    use crate::{prelude_test::Secret, schedule::MembershipKey};
+    use crate::{
+        framing::mls_auth_content::MlsAuthContent, prelude_test::Secret, schedule::MembershipKey,
+    };
 
     let epoch = random_u64();
     group.context_mut().set_epoch(epoch.into());
